pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "credit-risk-platform"
version = "0.1.0"
description = "Production-style credit risk modeling and decision API"
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Credit Risk Platform" }]
dependencies = [
  "boto3>=1.34.0",
  "fastapi>=0.110.0",
  "joblib>=1.4.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "pydantic>=2.6.0",
  "pyyaml>=6.0.0",
  "scikit-learn>=1.4.0",
  "scipy>=1.11.0",
  "uvicorn>=0.29.0",
]

[project.optional-dependencies]
aws = ["mangum>=0.17.0"]
explainability = ["shap>=0.46.0"]
dev = [
  "httpx>=0.27.0",
  "mypy>=1.9.0",
  "pytest>=8.1.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.3.0",
]

[project.scripts]
crp-train = "credit_risk.models.train:main"
crp-api = "credit_risk.api.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
